pyproject.toml
[project]
name = "march-18-dsrp-workshop"
version = "0.1.0"
description = "MLOps Workshop: Champion/Challenger con MLflow, DagsHub y GitHub Actions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "scikit-learn>=1.5.0",
    "mlflow>=2.12.0",
    "dagshub>=0.3.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "ipykernel>=6.29.0",
    "python-dotenv>=1.0.0",
]
